requirements.txt
alembic==1.13.1
APScheduler==3.10.4
async-timeout==4.0.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Migrate==4.0.7
Flask-Modals==0.5.1
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
reactivex==4.0.4
redis==5.0.6
RPi.GPIO==0.7.1
rq==1.16.2
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.0.3
WTForms==3.1.2
